Bugfixes regarding recurrence selection
